Finished major asset migration, swear filter
Few major changes Rayzr's previous formatting, but storage for player
and chunk data is now completely in the new, slower format.
However, chests are still dependent on Firebalance (worse yet, in the
old format) and miscellaneous data is stored in the misc file.
The swear filter automatically replaces anything in the blacklist
(including a dummy word, '%swear', for testing) with red asterisks, and
continues with the old color. That is all it does, so as to stay out of
the way.